refactor(admin): Refactor Swagger-related code
- Add SwaggerVersion enum import in SwaggerDocParser
- Update test case descriptions in SwaggerImportServiceTest
- Move SwaggerVersion enum from the admin module to the common module
